Change "touchstart" event handling
I have found in my applications that when I have a link or button within the scrollable div that `event.preventDefault` disables these. However, if I instead make it `event.stopPropagation` then the links and buttons work but the event is still not bubbled up above the element that you want scrolled.